Added some status messages in background fsp computation
Running on some testing data of much deeper sample sizes
than this program was originally developed for, with a
large number of different observed sample depths, the
background fsp computation took over an hour with the
program appearing to do nothing, so I added status messages
to indicate the program's progress during this step. This
part could be multithreaded to speed it up. Previously it
wasn't a significant concern.