Parallelized background fsp computation
Used openmp instead of pthreads. Actual selection scan is already
parallelized from back when this program was written in 2008, but
openmp wasn't available then or I didn't know about it. It is
still using pthreads and might as well. The background fsp stuff
is very easy to parallelize with openmp so going that route here.