Added cmdline option to specify minimum observation depth for SNPs
For large population samples, unimputed data may have a very large
number of different depths of observation resulting in the program
attempting to compute background fs and sweep model splines for
all of them, but a large fraction of these different sample depths
may have few SNPs associated with them. For current trial runs
using a sample with maximum depth >5400, and >2200 distinct sample
depths observed, restricting the minimum sample depth of SNPs to
as high as 5306 resulted in only a 2% loss in data, but >85% of
models no longer needed to be computed, saving a lot of memory
(like 100s of GB) and a lot of the pre-computation time for
computing the sweep models.

This new option allows someone to restrict the input that is loaded
to a minimum sample depth they have determined to have acceptable
loss in data for the gain in runtime efficiency and resource needs.
A better option in the future would determine a restriction
automatically. The current advantage is the input file does not
need to be filtered externally and the fscl program can be run with
different settings for this option on the same input file.