Disabled spline estimation of pjh[][] array. Added some README.md text.
There is something wrong with the spline estimation of pjh[][] which
is used for deep sample depths in place of the exact calculation. The
exact calculation actually doesn't take that much longer, but whether
or not it is more-than-you-can-tolerate depends on how many different
sample depths are observed, because it needs to be done for each of
them. It is an n^3 loop for the exact calculation. Normally spline
estimation is triggered at a sample depth of 800 or more. This is now
completely disabled and the exact calculation will always be used,
until I figure out what is wrong and fix it.

The program runs tolerably fast for sample depths of max=5630 as long
as the number of different sample depths is not huge. Dropping 10%
of sites with the highest amounts of missing data for the sample I
am playing with reduced the number of distinct sample depths by a
huge amount, allowing the program to proceed rather quickly to the
scanning and permutation stage in a few hours, using the exact
calculation.