Added implementation of using the non-central chisq distribution to estimate p-values.
Some other minor changes to improve multithreading efficiency. Also, the CLR being
recorded was actually the raw CLR and not twice the CLR as would be the test
statistic for normal likelihood-ratio tests (ie, not "composite" likelihood).
Switched this to be 2 * CLR, so the value if a normal likelihood ratio test would
have a chi-square distribution under the null hypothesis. Because this is
composite likelihood due to the non-independence between sites, the deviance from
this null distribution is expected to be a non-zero value for the non-centrality
parameter of the more general non-central chi-square distribution. The purpose
of this branch and change is to estimate that non-centrality parameter and then
use this distribution to estimate a p-value for the observed CLR.