Added CUDA code for calculating pjh[][] and pbk[][]
Focusing first on speeding up the pre-computation of the sweep models,
which can be lengthly, even days, in the case of a wide range of observed
sample depths, since a separate sweep model is computed for each one.

pbk[][] computation code is implemented but not tested. pjh[][] is tested
and can be successfully utilized by the program for 180X speedup (vs. 1 CPU)
w/ one Tesla K40 GPU. Code is not yet capable of using multiple GPUs in the
same system. The code to use the CUDA implementation for pjh[][] and pbk[][]
computation from sm-spline.c is not included in this commit.