Spline points are now calculated by CUDA GPU code. CPU not tested.
This commit not tested to work if -DCUDA is not used!

Added to pjh.cu code to calculate the points used to then fit
splines for the sweep models. The spline fitting itself is still
done by the original CPU code though this is an obvious next
place to convert to GPU.

The unit tests in pjh.cu verify the accuracy and correctness of the
GPU code vs. CPU equivalent and the numbers make sense. However, the
CPU code used for comparison is ported/copied from sm-spline.c to
pjh.cu and not the actual code that would be used if CUDA was not
defined at compile time. What would be used is the original CPU
code in sm-spline.c that has been modified some to accommodate
optionally using the GPU code, but I have not verified that this
mod'd code is still working if -DCUDA is not given to the command
line for the compiler. Ideally, the CPU code used when GPU is not
used should be that now in pjh.cu so that the unit tests there
for the GPU code are actually comparing against exactly what the
CPU version of the program does rather than a replica.