Added a simple R script for generating qq-plot with conf bounds
The script reads the fscl output file and generates the -log10(p)
qq-plot which should show a straight line for all but the very
highest points where presumably either sweeps have occurred at
those locations, or the permutation has not run far enough to
push these points to or over the line. Confidence bounds are
drawn. Thanks to Pavel Korniliev for sharing the R code to
generate these bounds.